Preserve both the wrapped-legend regression contracts and the new main branch contracts. Retain the legend sizing fix without further visual changes.

+++ internal/server/search_inspection_gradethis_integration_test.go
@@ -867,11 +867,9 @@ func gradeThisInspectionRouteLogicalPlan(
 		Stages:           stages,
 		ReferencedFields: slices.Clone(logical.ReferencedFields),
 		Output: &opensplunk.SearchInspectionOutputShape{
-			Kind:   gradeThisInspectionRouteOutputKind(logical.Output.Kind),
-			Fields: slices.Clone(logical.Output.Fields),
-			MaxDynamicFields: uint32(
-				logical.Output.MaxDynamicFields,
-			),
+			Kind:             gradeThisInspectionRouteOutputKind(logical.Output.Kind),
+			Fields:           slices.Clone(logical.Output.Fields),
+			MaxDynamicFields: logical.Output.MaxDynamicFields,
 		},
 	}
 }
